+++ src/services/WidgetService.js
@@ -17,40 +17,120 @@ export default class WidgetService {
       .then(jsonData => jsonData.data);
   }
 
-  static saveUserWidget(widget, datasetId, token) {
-    const widgetObj = {
+  /**
+   * Create a widget in the API and its associated metadata (optional)
+   * @param {string} datasetId ID of the dataset
+   * @param {string} token Token of the user
+   * @param {object} widgetBody Content of request to create the widget
+   * @param {object} metadataBody Content of the request to create the metadata
+   */
+  static saveUserWidget(datasetId, token, widgetBody, metadataBody = null) {
+    const widget = Object.assign({}, widgetBody, {
       application: [getConfig().applications],
       published: false,
       default: false,
       dataset: datasetId
-    };
-    const bodyObj = Object.assign({}, widget, widgetObj);
-    return fetch(`${getConfig().url}/dataset/${datasetId}/widget`, {
+    });
+
+    const metadata = !metadataBody
+      ? null
+      : Object.assign({}, metadataBody, {
+        language: getConfig().locale,
+        application: getConfig().applications
+      });
+
+    const getRequestOptions = body => ({
       method: 'POST',
-      body: JSON.stringify(bodyObj),
+      body: JSON.stringify(body),
       headers: {
         'Content-Type': 'application/json',
         Authorization: `Bearer ${token}`
       }
-    })
-      .then((response) => {
-        if (response.status >= 400) throw new Error(response.statusText);
-        return response.json();
+    });
+
+    return fetch(`${getConfig().url}/dataset/${datasetId}/widget`, getRequestOptions(widget))
+      .then((res) => {
+        if (!res.ok) {
+          console.error('Unable to create the widget');
+          throw new Error(res.statusText);
+        }
+        return res.json();
+      })
+      .then((data) => {
+        if (metadata) {
+          return fetch(`${getConfig().url}/dataset/${datasetId}/widget/${data.data.id}/metadata`, getRequestOptions(metadata));
+        }
+        return data;
+      })
+      .then((res) => {
+        if (metadata) {
+          if (!res.ok) {
+            console.error('Unable to create the metadata');
+            throw new Error(res.statusText);
+          }
+          return res.json();
+        }
+
+        return res;
       });
   }
 
-  static updateUserWidget(widget, datasetId, token) {
-    return fetch(`${getConfig().url}/dataset/${datasetId}/widget/${widget.id}`, {
+  /**
+   * Update a widget of the API and its associated metadata (optional)
+   * @param {string} datasetId ID of the dataset
+   * @param {string} widgetId ID of the dataset
+   * @param {string} token Token of the user
+   * @param {object} widgetBody Content of request to create the widget
+   * @param {object} metadataBody Content of the request to create the metadata
+   */
+  static updateUserWidget(datasetId, widgetId, token, widgetBody, metadataBody = null) {
+    const widget = Object.assign({}, widgetBody, {
+      application: [getConfig().applications],
+      published: false,
+      default: false,
+      dataset: datasetId
+    });
+
+    const metadata = !metadataBody
+      ? null
+      : Object.assign({}, metadataBody, {
+        language: getConfig().locale,
+        application: getConfig().applications
+      });
+
+    const getRequestOptions = body => ({
       method: 'PATCH',
-      body: JSON.stringify(widget),
+      body: JSON.stringify(body),
       headers: {
         'Content-Type': 'application/json',
         Authorization: `Bearer ${token}`
       }
-    })
-      .then((response) => {
-        if (response.status >= 400) throw new Error(response.statusText);
-        return response.json();
+    });
+
+    return fetch(`${getConfig().url}/dataset/${datasetId}/widget/${widgetId}`, getRequestOptions(widget))
+      .then((res) => {
+        if (!res.ok) {
+          console.error('Unable to update the widget');
+          throw new Error(res.statusText);
+        }
+        return res.json();
+      })
+      .then((data) => {
+        if (metadata) {
+          return fetch(`${getConfig().url}/dataset/${datasetId}/widget/${widgetId}/metadata`, getRequestOptions(metadata));
+        }
+        return data;
+      })
+      .then((res) => {
+        if (metadata) {
+          if (!res.ok) {
+            console.error('Unable to update the metadata');
+            throw new Error(res.statusText);
+          }
+          return res.json();
+        }
+
+        return res;
       });
   }