+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Add a new prop, `theme`, to customize the visual appearance of the Vega visualizations
 - Remove Leaflet from the peer dependencies (possibility to load it from a CDN)
 - Fix a bug where the aggregation would often not be restored
+- Change the signature of the `saveUserWidget` and `updateUserWidget` functions of [`WidgetService`](https://github.com/resource-watch/widget-editor/blob/develop/src/services/WidgetService.js)
+- Save and restore the caption from the metadata of the widget instead of the `widgetConfig` object
 
 ## v1.0.6 - 10/04/2018
 - Improve the styles of the columns, especially when overflowing